requirements.txt
python-telegram-bot==21.10
openai>=1.0.0
python-dotenv
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
